Refactor authentication form component
