+++ src/components/AuthForm.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useRef } from 'react';
+import { useState, useEffect, useRef } from 'react';
 import Grid from '@mui/material/Grid';
 import Box from '@mui/material/Box';
 import TextField from '@mui/material/TextField';
@@ -6,20 +6,29 @@ import Button from '@mui/material/Button';
 
 type AuthType = 'login' | 'signup';
 
-interface Props {
-  type: AuthType;
-  switchType: () => void;
-  handleSubmit: React.FormEventHandler<HTMLFormElement>;
-}
+const AuthForm = () => {
+  const [authType, setAuthType] = useState<AuthType>('login');
 
-const AuthForm = ({ type, switchType, handleSubmit }: Props) => {
   const nameFieldRef = useRef<HTMLInputElement>(null);
   const mailFieldRef = useRef<HTMLInputElement>(null);
 
   useEffect(() => {
-    const fieldRef = type === 'login' ? mailFieldRef : nameFieldRef;
+    const fieldRef = authType === 'login' ? mailFieldRef : nameFieldRef;
     fieldRef.current?.focus();
-  }, [type]);
+  }, [authType]);
+
+  const handleLogin = (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+  };
+
+  const handleSignup = (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+  };
+
+  const handlers = {
+    login: handleLogin,
+    signup: handleSignup,
+  };
 
   return (
     <Box
@@ -33,41 +42,39 @@ const AuthForm = ({ type, switchType, handleSubmit }: Props) => {
         <Button
           variant="text"
           color="inherit"
-          onClick={switchType}
-          disabled={type === 'login'}
+          onClick={() => setAuthType('login')}
+          disabled={authType === 'login'}
         >
           Log In
         </Button>
         <Button
           variant="text"
           color="inherit"
-          onClick={switchType}
-          disabled={type === 'signup'}
+          onClick={() => setAuthType('signup')}
+          disabled={authType === 'signup'}
         >
           Sign Up
         </Button>
       </Box>
-      <Box component="form" maxWidth="sm" onSubmit={handleSubmit}>
+      <Box component="form" maxWidth="sm" onSubmit={handlers[authType]}>
         <Grid container spacing={2}>
-          <Grid item xs={12} sm={6}>
+          <Grid item xs={12} sm={6} hidden={authType === 'login'}>
             <TextField
               id="first-name"
               name="first-name"
               label="First name"
               inputRef={nameFieldRef}
               autoComplete="given-name"
-              disabled={type === 'login'}
               required
               fullWidth
             />
           </Grid>
-          <Grid item xs={12} sm={6}>
+          <Grid item xs={12} sm={6} hidden={authType === 'login'}>
             <TextField
               id="last-name"
               name="last-name"
               label="Last name"
               autoComplete="family-name"
-              disabled={type === 'login'}
               required
               fullWidth
             />
@@ -96,7 +103,7 @@ const AuthForm = ({ type, switchType, handleSubmit }: Props) => {
           </Grid>
           <Grid item xs={12}>
             <Button type="submit" variant="contained" size="large" fullWidth>
-              {type === 'login' ? 'Log In' : 'Sign Up'}
+              {authType === 'login' ? 'Log In' : 'Sign Up'}
             </Button>
           </Grid>
         </Grid>
@@ -106,4 +113,3 @@ const AuthForm = ({ type, switchType, handleSubmit }: Props) => {
 };
 
 export default AuthForm;
-export type { AuthType };
